backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
transformers==4.41.2
torch==2.3.1
accelerate==0.31.0
sentencepiece==0.2.0
protobuf==5.27.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
